clean up authentication tests to their own test class
